Network.AWS.ECR

Description

Amazon EC2 Container Registry (Amazon ECR) is a managed AWS Docker registry service. Customers can use the familiar Docker CLI to push, pull, and manage images. Amazon ECR provides a secure, scalable, and reliable registry. Amazon ECR supports private Docker repositories with resource-based permissions using AWS IAM so that specific users or Amazon EC2 instances can access repositories and images. Developers can use the Docker CLI to author and manage images.

Service Configuration

ecr :: Service

API version 2015-09-21 of the Amazon EC2 Container Registry SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by ECR.

LayersNotFoundException

_LayersNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified layers could not be found, or the specified layer is not valid for this repository.

InvalidParameterException

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

The specified parameter is invalid. Review the available parameters for the API request.

LayerAlreadyExistsException

_LayerAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The image layer already exists in the associated repository.

ServerException

_ServerException :: AsError a => Getting (First ServiceError) a ServiceError

These errors are usually caused by a server-side issue.

LayerInaccessibleException

_LayerInaccessibleException :: AsError a => Getting (First ServiceError) a ServiceError

The specified layer is not available because it is not associated with an image. Unassociated image layers may be cleaned up at any time.

InvalidLayerException

_InvalidLayerException :: AsError a => Getting (First ServiceError) a ServiceError

The layer digest calculation performed by Amazon ECR upon receipt of the image layer does not match the digest specified.

LayerPartTooSmallException

_LayerPartTooSmallException :: AsError a => Getting (First ServiceError) a ServiceError

Layer parts must be at least 5 MiB in size.

ImageNotFoundException

_ImageNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The image requested does not exist in the specified repository.

ImageAlreadyExistsException

_ImageAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The specified image has already been pushed, and there are no changes to the manifest or image tag since the last push.

RepositoryNotFoundException

_RepositoryNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified repository could not be found. Check the spelling of the specified repository and ensure that you are performing operations on the correct registry.

UploadNotFoundException

_UploadNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The upload could not be found, or the specified upload id is not valid for this repository.

InvalidLayerPartException

_InvalidLayerPartException :: AsError a => Getting (First ServiceError) a ServiceError

The layer part size is not valid, or the first byte specified is not consecutive to the last byte of a previous layer part upload.

RepositoryNotEmptyException

_RepositoryNotEmptyException :: AsError a => Getting (First ServiceError) a ServiceError

The specified repository contains images. To delete a repository that contains images, you must force the deletion with the force parameter.

RepositoryAlreadyExistsException

_RepositoryAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The specified repository already exists in the specified registry.

RepositoryPolicyNotFoundException

_RepositoryPolicyNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified repository and registry combination does not have an associated repository policy.

EmptyUploadException

_EmptyUploadException :: AsError a => Getting (First ServiceError) a ServiceError

The specified layer upload does not contain any layer parts.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The operation did not succeed because it would have exceeded a service limit for your account. For more information, see Amazon ECR Default Service Limits in the Amazon EC2 Container Registry User Guide.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetRepositoryPolicy

module Network.AWS.ECR.GetRepositoryPolicy

BatchDeleteImage

module Network.AWS.ECR.BatchDeleteImage

BatchCheckLayerAvailability

module Network.AWS.ECR.BatchCheckLayerAvailability

DeleteRepositoryPolicy

module Network.AWS.ECR.DeleteRepositoryPolicy

CreateRepository

module Network.AWS.ECR.CreateRepository

CompleteLayerUpload

module Network.AWS.ECR.CompleteLayerUpload

DescribeRepositories (Paginated)

module Network.AWS.ECR.DescribeRepositories

UploadLayerPart

module Network.AWS.ECR.UploadLayerPart

BatchGetImage

module Network.AWS.ECR.BatchGetImage

SetRepositoryPolicy

module Network.AWS.ECR.SetRepositoryPolicy

InitiateLayerUpload

module Network.AWS.ECR.InitiateLayerUpload

DeleteRepository

module Network.AWS.ECR.DeleteRepository

PutImage

module Network.AWS.ECR.PutImage

ListImages (Paginated)

module Network.AWS.ECR.ListImages

GetAuthorizationToken

module Network.AWS.ECR.GetAuthorizationToken

GetDownloadURLForLayer

module Network.AWS.ECR.GetDownloadURLForLayer

DescribeImages (Paginated)

module Network.AWS.ECR.DescribeImages

Types

ImageFailureCode

data ImageFailureCode

Constructors

ImageNotFound

ImageTagDoesNotMatchDigest

InvalidImageDigest

InvalidImageTag

MissingDigestAndTag

LayerAvailability

data LayerAvailability

Constructors

Available

Unavailable

LayerFailureCode

data LayerFailureCode

Constructors

InvalidLayerDigest

MissingLayerDigest

TagStatus

data TagStatus

Constructors

Tagged

Untagged

AuthorizationData

data AuthorizationData

An object representing authorization data for an Amazon ECR registry.

See: authorizationData smart constructor.

authorizationData :: AuthorizationData

Creates a value of AuthorizationData with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

adExpiresAt - The Unix time in seconds and milliseconds when the authorization token expires. Authorization tokens are valid for 12 hours.

adProxyEndpoint - The registry URL to use for this authorization token in a docker login command. The Amazon ECR registry URL format is https://aws_account_id.dkr.ecr.region.amazonaws.com . For example, https://012345678910.dkr.ecr.us-east-1.amazonaws.com ..

adAuthorizationToken - A base64-encoded string that contains authorization data for the specified Amazon ECR registry. When the string is decoded, it is presented in the format user:password for private registry authentication using docker login .

adExpiresAt :: Lens' AuthorizationData (Maybe UTCTime)

The Unix time in seconds and milliseconds when the authorization token expires. Authorization tokens are valid for 12 hours.

adProxyEndpoint :: Lens' AuthorizationData (Maybe Text)

The registry URL to use for this authorization token in a docker login command. The Amazon ECR registry URL format is https://aws_account_id.dkr.ecr.region.amazonaws.com . For example, https://012345678910.dkr.ecr.us-east-1.amazonaws.com ..

adAuthorizationToken :: Lens' AuthorizationData (Maybe Text)

A base64-encoded string that contains authorization data for the specified Amazon ECR registry. When the string is decoded, it is presented in the format user:password for private registry authentication using docker login .

DescribeImagesFilter

data DescribeImagesFilter

An object representing a filter on a DescribeImages operation.

See: describeImagesFilter smart constructor.

describeImagesFilter :: DescribeImagesFilter

Creates a value of DescribeImagesFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

difTagStatus - The tag status with which to filter your DescribeImages results. You can filter results based on whether they are TAGGED or UNTAGGED .

difTagStatus :: Lens' DescribeImagesFilter (Maybe TagStatus)

The tag status with which to filter your DescribeImages results. You can filter results based on whether they are TAGGED or UNTAGGED .

Image

data Image

An object representing an Amazon ECR image.

See: image smart constructor.

image :: Image

Creates a value of Image with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iRegistryId - The AWS account ID associated with the registry containing the image.

iImageId - An object containing the image tag and image digest associated with an image.

iRepositoryName - The name of the repository associated with the image.

iImageManifest - The image manifest associated with the image.

iRegistryId :: Lens' Image (Maybe Text)

The AWS account ID associated with the registry containing the image.

iImageId :: Lens' Image (Maybe ImageIdentifier)

An object containing the image tag and image digest associated with an image.

iRepositoryName :: Lens' Image (Maybe Text)

The name of the repository associated with the image.

iImageManifest :: Lens' Image (Maybe Text)

The image manifest associated with the image.

ImageDetail

data ImageDetail

An object that describes an image returned by a DescribeImages operation.

See: imageDetail smart constructor.

imageDetail :: ImageDetail

Creates a value of ImageDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

idRegistryId - The AWS account ID associated with the registry to which this image belongs.

idImageTags - The list of tags associated with this image.

idImageSizeInBytes - The size, in bytes, of the image in the repository.

idImageDigest - The sha256 digest of the image manifest.

idImagePushedAt - The date and time, expressed in standard JavaScript date format, at which the current image was pushed to the repository.

idRepositoryName - The name of the repository to which this image belongs.

idRegistryId :: Lens' ImageDetail (Maybe Text)

The AWS account ID associated with the registry to which this image belongs.

idImageTags :: Lens' ImageDetail [Text]

The list of tags associated with this image.

idImageSizeInBytes :: Lens' ImageDetail (Maybe Integer)

The size, in bytes, of the image in the repository.

idImageDigest :: Lens' ImageDetail (Maybe Text)

The sha256 digest of the image manifest.

idImagePushedAt :: Lens' ImageDetail (Maybe UTCTime)

The date and time, expressed in standard JavaScript date format, at which the current image was pushed to the repository.

idRepositoryName :: Lens' ImageDetail (Maybe Text)

The name of the repository to which this image belongs.

ImageFailure

data ImageFailure

An object representing an Amazon ECR image failure.

See: imageFailure smart constructor.

imageFailure :: ImageFailure

Creates a value of ImageFailure with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ifFailureReason - The reason for the failure.

ifFailureCode - The code associated with the failure.

ifImageId - The image ID associated with the failure.

ifFailureReason :: Lens' ImageFailure (Maybe Text)

The reason for the failure.

ifFailureCode :: Lens' ImageFailure (Maybe ImageFailureCode)

The code associated with the failure.

ifImageId :: Lens' ImageFailure (Maybe ImageIdentifier)

The image ID associated with the failure.

ImageIdentifier

data ImageIdentifier

An object with identifying information for an Amazon ECR image.

See: imageIdentifier smart constructor.

imageIdentifier :: ImageIdentifier

Creates a value of ImageIdentifier with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iiImageDigest - The sha256 digest of the image manifest.

iiImageTag - The tag used for the image.

iiImageDigest :: Lens' ImageIdentifier (Maybe Text)

The sha256 digest of the image manifest.

iiImageTag :: Lens' ImageIdentifier (Maybe Text)

The tag used for the image.

Layer

data Layer

An object representing an Amazon ECR image layer.

See: layer smart constructor.

layer :: Layer

Creates a value of Layer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lMediaType - The media type of the layer, such as application/vnd.docker.image.rootfs.diff.tar.gzip or application/vnd.oci.image.layer.v1.tar+gzip .

lLayerDigest - The sha256 digest of the image layer.

lLayerSize - The size, in bytes, of the image layer.

lLayerAvailability - The availability status of the image layer.

lMediaType :: Lens' Layer (Maybe Text)

The media type of the layer, such as application/vnd.docker.image.rootfs.diff.tar.gzip or application/vnd.oci.image.layer.v1.tar+gzip .

lLayerDigest :: Lens' Layer (Maybe Text)

The sha256 digest of the image layer.

lLayerSize :: Lens' Layer (Maybe Integer)

The size, in bytes, of the image layer.

lLayerAvailability :: Lens' Layer (Maybe LayerAvailability)

The availability status of the image layer.

LayerFailure

data LayerFailure

An object representing an Amazon ECR image layer failure.

See: layerFailure smart constructor.

layerFailure :: LayerFailure

Creates a value of LayerFailure with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lfFailureReason - The reason for the failure.

lfFailureCode - The failure code associated with the failure.

lfLayerDigest - The layer digest associated with the failure.

lfFailureReason :: Lens' LayerFailure (Maybe Text)

The reason for the failure.

lfFailureCode :: Lens' LayerFailure (Maybe LayerFailureCode)

The failure code associated with the failure.

lfLayerDigest :: Lens' LayerFailure (Maybe Text)

The layer digest associated with the failure.

ListImagesFilter

data ListImagesFilter

An object representing a filter on a ListImages operation.

See: listImagesFilter smart constructor.

listImagesFilter :: ListImagesFilter

Creates a value of ListImagesFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lifTagStatus - The tag status with which to filter your ListImages results. You can filter results based on whether they are TAGGED or UNTAGGED .

lifTagStatus :: Lens' ListImagesFilter (Maybe TagStatus)

The tag status with which to filter your ListImages results. You can filter results based on whether they are TAGGED or UNTAGGED .

Repository

data Repository

An object representing a repository.

See: repository smart constructor.

repository :: Repository

Creates a value of Repository with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rRepositoryARN - The Amazon Resource Name (ARN) that identifies the repository. The ARN contains the arn:aws:ecr namespace, followed by the region of the repository, the AWS account ID of the repository owner, the repository namespace, and then the repository name. For example, arn:aws:ecr:region:012345678910:repository/test .

rCreatedAt - The date and time, in JavaScript date/time format, when the repository was created.

rRegistryId - The AWS account ID associated with the registry that contains the repository.

rRepositoryURI - The URI for the repository. You can use this URI for Docker push and pull operations.

rRepositoryName - The name of the repository.

rRepositoryARN :: Lens' Repository (Maybe Text)

The Amazon Resource Name (ARN) that identifies the repository. The ARN contains the arn:aws:ecr namespace, followed by the region of the repository, the AWS account ID of the repository owner, the repository namespace, and then the repository name. For example, arn:aws:ecr:region:012345678910:repository/test .

rCreatedAt :: Lens' Repository (Maybe UTCTime)

The date and time, in JavaScript date/time format, when the repository was created.

rRegistryId :: Lens' Repository (Maybe Text)

The AWS account ID associated with the registry that contains the repository.

rRepositoryURI :: Lens' Repository (Maybe Text)

The URI for the repository. You can use this URI for Docker push and pull operations.

rRepositoryName :: Lens' Repository (Maybe Text)

The name of the repository.