Network.AWS.ECR.BatchCheckLayerAvailability

Description

Check the availability of multiple image layers in a specified registry and repository.

Creating a Request

batchCheckLayerAvailability

Arguments

:: Text
bclaRepositoryName

-> NonEmpty Text
bclaLayerDigests

-> BatchCheckLayerAvailability

Creates a value of BatchCheckLayerAvailability with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bclaRegistryId - The AWS account ID associated with the registry that contains the image layers to check. If you do not specify a registry, the default registry is assumed.

bclaRepositoryName - The name of the repository that is associated with the image layers to check.

bclaLayerDigests - The digests of the image layers to check.

data BatchCheckLayerAvailability

See: batchCheckLayerAvailability smart constructor.

Request Lenses

bclaRegistryId :: Lens' BatchCheckLayerAvailability (Maybe Text)

The AWS account ID associated with the registry that contains the image layers to check. If you do not specify a registry, the default registry is assumed.

bclaRepositoryName :: Lens' BatchCheckLayerAvailability Text

The name of the repository that is associated with the image layers to check.

bclaLayerDigests :: Lens' BatchCheckLayerAvailability (NonEmpty Text)

The digests of the image layers to check.

Destructuring the Response

batchCheckLayerAvailabilityResponse

Arguments

:: Int
bclarsResponseStatus

-> BatchCheckLayerAvailabilityResponse

Creates a value of BatchCheckLayerAvailabilityResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bclarsFailures - Any failures associated with the call.

bclarsLayers - A list of image layer objects corresponding to the image layer references in the request.

bclarsResponseStatus - -- | The response status code.

data BatchCheckLayerAvailabilityResponse

See: batchCheckLayerAvailabilityResponse smart constructor.

Response Lenses

bclarsFailures :: Lens' BatchCheckLayerAvailabilityResponse [LayerFailure]

Any failures associated with the call.

bclarsLayers :: Lens' BatchCheckLayerAvailabilityResponse [Layer]

A list of image layer objects corresponding to the image layer references in the request.

bclarsResponseStatus :: Lens' BatchCheckLayerAvailabilityResponse Int

- | The response status code.