Network.AWS.ECR.BatchGetImage

Description

Gets detailed information for specified images within a specified repository. Images are specified with either imageTag or imageDigest .

Creating a Request

batchGetImage

Arguments

:: Text
bgiRepositoryName

-> BatchGetImage

Creates a value of BatchGetImage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bgiRegistryId - The AWS account ID associated with the registry that contains the images to describe. If you do not specify a registry, the default registry is assumed.

bgiAcceptedMediaTypes - The accepted media types for the request. Valid values: application/vnd.docker.distribution.manifest.v1+json | application/vnd.docker.distribution.manifest.v2+json | application/vnd.oci.image.manifest.v1+json

bgiRepositoryName - The repository that contains the images to describe.

bgiImageIds - A list of image ID references that correspond to images to describe. The format of the imageIds reference is imageTag=tag or imageDigest=digest .

data BatchGetImage

See: batchGetImage smart constructor.

Request Lenses

bgiRegistryId :: Lens' BatchGetImage (Maybe Text)

The AWS account ID associated with the registry that contains the images to describe. If you do not specify a registry, the default registry is assumed.

bgiAcceptedMediaTypes :: Lens' BatchGetImage (Maybe (NonEmpty Text))

The accepted media types for the request. Valid values: application/vnd.docker.distribution.manifest.v1+json | application/vnd.docker.distribution.manifest.v2+json | application/vnd.oci.image.manifest.v1+json

bgiRepositoryName :: Lens' BatchGetImage Text

The repository that contains the images to describe.

bgiImageIds :: Lens' BatchGetImage [ImageIdentifier]

A list of image ID references that correspond to images to describe. The format of the imageIds reference is imageTag=tag or imageDigest=digest .

Destructuring the Response

batchGetImageResponse

Arguments

:: Int
bgirsResponseStatus

-> BatchGetImageResponse

Creates a value of BatchGetImageResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bgirsImages - A list of image objects corresponding to the image references in the request.

bgirsFailures - Any failures associated with the call.

bgirsResponseStatus - -- | The response status code.

data BatchGetImageResponse

See: batchGetImageResponse smart constructor.

Response Lenses

bgirsImages :: Lens' BatchGetImageResponse [Image]

A list of image objects corresponding to the image references in the request.

bgirsFailures :: Lens' BatchGetImageResponse [ImageFailure]

Any failures associated with the call.

bgirsResponseStatus :: Lens' BatchGetImageResponse Int

- | The response status code.