Network.AWS.ECR.DeleteRepository

Description

Deletes an existing image repository. If a repository contains images, you must use the force option to delete it.

Creating a Request

deleteRepository

Arguments

:: Text
dRepositoryName

-> DeleteRepository

Creates a value of DeleteRepository with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dForce - Force the deletion of the repository if it contains images.

dRegistryId - The AWS account ID associated with the registry that contains the repository to delete. If you do not specify a registry, the default registry is assumed.

dRepositoryName - The name of the repository to delete.

data DeleteRepository

See: deleteRepository smart constructor.

Request Lenses

dForce :: Lens' DeleteRepository (Maybe Bool)

Force the deletion of the repository if it contains images.

dRegistryId :: Lens' DeleteRepository (Maybe Text)

The AWS account ID associated with the registry that contains the repository to delete. If you do not specify a registry, the default registry is assumed.

dRepositoryName :: Lens' DeleteRepository Text

The name of the repository to delete.

Destructuring the Response

deleteRepositoryResponse

Arguments

:: Int
drsResponseStatus

-> DeleteRepositoryResponse

Creates a value of DeleteRepositoryResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsRepository - The repository that was deleted.

drsResponseStatus - -- | The response status code.

data DeleteRepositoryResponse

See: deleteRepositoryResponse smart constructor.

Response Lenses

drsRepository :: Lens' DeleteRepositoryResponse (Maybe Repository)

The repository that was deleted.

drsResponseStatus :: Lens' DeleteRepositoryResponse Int

- | The response status code.