Network.AWS.ECR.DescribeImages

Description

Returns metadata about the images in a repository, including image size, image tags, and creation date.

This operation returns paginated results.

Creating a Request

describeImages

Arguments

:: Text
diRepositoryName

-> DescribeImages

Creates a value of DescribeImages with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diRegistryId - The AWS account ID associated with the registry that contains the repository in which to describe images. If you do not specify a registry, the default registry is assumed.

diImageIds - The list of image IDs for the requested repository.

diNextToken - The nextToken value returned from a previous paginated DescribeImages request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

diFilter - The filter key and value with which to filter your DescribeImages results.

diMaxResults - The maximum number of repository results returned by DescribeImages in paginated output. When this parameter is used, DescribeImages only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeImages request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then DescribeImages returns up to 100 results and a nextToken value, if applicable.

diRepositoryName - A list of repositories to describe. If this parameter is omitted, then all repositories in a registry are described.

data DescribeImages

See: describeImages smart constructor.

Request Lenses

diRegistryId :: Lens' DescribeImages (Maybe Text)

The AWS account ID associated with the registry that contains the repository in which to describe images. If you do not specify a registry, the default registry is assumed.

diImageIds :: Lens' DescribeImages [ImageIdentifier]

The list of image IDs for the requested repository.

diNextToken :: Lens' DescribeImages (Maybe Text)

The nextToken value returned from a previous paginated DescribeImages request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

diFilter :: Lens' DescribeImages (Maybe DescribeImagesFilter)

The filter key and value with which to filter your DescribeImages results.

diMaxResults :: Lens' DescribeImages (Maybe Natural)

The maximum number of repository results returned by DescribeImages in paginated output. When this parameter is used, DescribeImages only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeImages request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then DescribeImages returns up to 100 results and a nextToken value, if applicable.

diRepositoryName :: Lens' DescribeImages Text

A list of repositories to describe. If this parameter is omitted, then all repositories in a registry are described.

Destructuring the Response

describeImagesResponse

Arguments

:: Int
dirsResponseStatus

-> DescribeImagesResponse

Creates a value of DescribeImagesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dirsImageDetails - A list of ImageDetail objects that contain data about the image.

dirsNextToken - The nextToken value to include in a future DescribeImages request. When the results of a DescribeImages request exceed maxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

dirsResponseStatus - -- | The response status code.

data DescribeImagesResponse

See: describeImagesResponse smart constructor.

Response Lenses

dirsImageDetails :: Lens' DescribeImagesResponse [ImageDetail]

A list of ImageDetail objects that contain data about the image.

dirsNextToken :: Lens' DescribeImagesResponse (Maybe Text)

The nextToken value to include in a future DescribeImages request. When the results of a DescribeImages request exceed maxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

dirsResponseStatus :: Lens' DescribeImagesResponse Int

- | The response status code.