Network.AWS.ECR.DescribeRepositories

Description

Describes image repositories in a registry.

This operation returns paginated results.

Creating a Request

describeRepositories :: DescribeRepositories

Creates a value of DescribeRepositories with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drRegistryId - The AWS account ID associated with the registry that contains the repositories to be described. If you do not specify a registry, the default registry is assumed.

drRepositoryNames - A list of repositories to describe. If this parameter is omitted, then all repositories in a registry are described.

drNextToken - The nextToken value returned from a previous paginated DescribeRepositories request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

drMaxResults - The maximum number of repository results returned by DescribeRepositories in paginated output. When this parameter is used, DescribeRepositories only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeRepositories request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then DescribeRepositories returns up to 100 results and a nextToken value, if applicable.

data DescribeRepositories

See: describeRepositories smart constructor.

Request Lenses

drRegistryId :: Lens' DescribeRepositories (Maybe Text)

The AWS account ID associated with the registry that contains the repositories to be described. If you do not specify a registry, the default registry is assumed.

drRepositoryNames :: Lens' DescribeRepositories (Maybe (NonEmpty Text))

A list of repositories to describe. If this parameter is omitted, then all repositories in a registry are described.

drNextToken :: Lens' DescribeRepositories (Maybe Text)

The nextToken value returned from a previous paginated DescribeRepositories request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

drMaxResults :: Lens' DescribeRepositories (Maybe Natural)

The maximum number of repository results returned by DescribeRepositories in paginated output. When this parameter is used, DescribeRepositories only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeRepositories request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then DescribeRepositories returns up to 100 results and a nextToken value, if applicable.

Destructuring the Response

describeRepositoriesResponse

Arguments

:: Int
drrsResponseStatus

-> DescribeRepositoriesResponse

Creates a value of DescribeRepositoriesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drrsRepositories - A list of repository objects corresponding to valid repositories.

drrsNextToken - The nextToken value to include in a future DescribeRepositories request. When the results of a DescribeRepositories request exceed maxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

drrsResponseStatus - -- | The response status code.

data DescribeRepositoriesResponse

See: describeRepositoriesResponse smart constructor.

Response Lenses

drrsRepositories :: Lens' DescribeRepositoriesResponse [Repository]

A list of repository objects corresponding to valid repositories.

drrsNextToken :: Lens' DescribeRepositoriesResponse (Maybe Text)

The nextToken value to include in a future DescribeRepositories request. When the results of a DescribeRepositories request exceed maxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

drrsResponseStatus :: Lens' DescribeRepositoriesResponse Int

- | The response status code.