Network.AWS.ECR.GetAuthorizationToken

Description

Retrieves a token that is valid for a specified registry for 12 hours. This command allows you to use the docker CLI to push and pull images with Amazon ECR. If you do not specify a registry, the default registry is assumed.

The authorizationToken returned for each registry specified is a base64 encoded string that can be decoded and used in a docker login command to authenticate to a registry. The AWS CLI offers an aws ecr get-login command that simplifies the login process.

Creating a Request

getAuthorizationToken :: GetAuthorizationToken

Creates a value of GetAuthorizationToken with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gatRegistryIds - A list of AWS account IDs that are associated with the registries for which to get authorization tokens. If you do not specify a registry, the default registry is assumed.

data GetAuthorizationToken

See: getAuthorizationToken smart constructor.

Request Lenses

gatRegistryIds :: Lens' GetAuthorizationToken (Maybe (NonEmpty Text))

A list of AWS account IDs that are associated with the registries for which to get authorization tokens. If you do not specify a registry, the default registry is assumed.

Destructuring the Response

getAuthorizationTokenResponse

Arguments

:: Int
gatrsResponseStatus

-> GetAuthorizationTokenResponse

Creates a value of GetAuthorizationTokenResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gatrsAuthorizationData - A list of authorization token data objects that correspond to the registryIds values in the request.

gatrsResponseStatus - -- | The response status code.

data GetAuthorizationTokenResponse

See: getAuthorizationTokenResponse smart constructor.

Response Lenses

gatrsAuthorizationData :: Lens' GetAuthorizationTokenResponse [AuthorizationData]

A list of authorization token data objects that correspond to the registryIds values in the request.

gatrsResponseStatus :: Lens' GetAuthorizationTokenResponse Int

- | The response status code.