Network.AWS.ECR.GetRepositoryPolicy

Description

Retrieves the repository policy for a specified repository.

Creating a Request

getRepositoryPolicy

Arguments

:: Text
grpRepositoryName

-> GetRepositoryPolicy

Creates a value of GetRepositoryPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grpRegistryId - The AWS account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

grpRepositoryName - The name of the repository whose policy you want to retrieve.

data GetRepositoryPolicy

See: getRepositoryPolicy smart constructor.

Request Lenses

grpRegistryId :: Lens' GetRepositoryPolicy (Maybe Text)

The AWS account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

grpRepositoryName :: Lens' GetRepositoryPolicy Text

The name of the repository whose policy you want to retrieve.

Destructuring the Response

getRepositoryPolicyResponse

Arguments

:: Int
grprsResponseStatus

-> GetRepositoryPolicyResponse

Creates a value of GetRepositoryPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grprsRegistryId - The registry ID associated with the request.

grprsRepositoryName - The repository name associated with the request.

grprsPolicyText - The JSON repository policy text associated with the repository.

grprsResponseStatus - -- | The response status code.

data GetRepositoryPolicyResponse

See: getRepositoryPolicyResponse smart constructor.

Response Lenses

grprsRegistryId :: Lens' GetRepositoryPolicyResponse (Maybe Text)

The registry ID associated with the request.

grprsRepositoryName :: Lens' GetRepositoryPolicyResponse (Maybe Text)

The repository name associated with the request.

grprsPolicyText :: Lens' GetRepositoryPolicyResponse (Maybe Text)

The JSON repository policy text associated with the repository.

grprsResponseStatus :: Lens' GetRepositoryPolicyResponse Int

- | The response status code.