Network.AWS.ECR.ListImages

Description

Lists all the image IDs for a given repository.

You can filter images based on whether or not they are tagged by setting the tagStatus parameter to TAGGED or UNTAGGED . For example, you can filter your results to return only UNTAGGED images and then pipe that result to a BatchDeleteImage operation to delete them. Or, you can filter your results to return only TAGGED images to list all of the tags in your repository.

This operation returns paginated results.

Creating a Request

listImages

Arguments

:: Text
liRepositoryName

-> ListImages

Creates a value of ListImages with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

liRegistryId - The AWS account ID associated with the registry that contains the repository to list images in. If you do not specify a registry, the default registry is assumed.

liNextToken - The nextToken value returned from a previous paginated ListImages request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

liFilter - The filter key and value with which to filter your ListImages results.

liMaxResults - The maximum number of image results returned by ListImages in paginated output. When this parameter is used, ListImages only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListImages request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListImages returns up to 100 results and a nextToken value, if applicable.

liRepositoryName - The repository whose image IDs are to be listed.

data ListImages

See: listImages smart constructor.

Request Lenses

liRegistryId :: Lens' ListImages (Maybe Text)

The AWS account ID associated with the registry that contains the repository to list images in. If you do not specify a registry, the default registry is assumed.

liNextToken :: Lens' ListImages (Maybe Text)

The nextToken value returned from a previous paginated ListImages request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

liFilter :: Lens' ListImages (Maybe ListImagesFilter)

The filter key and value with which to filter your ListImages results.

liMaxResults :: Lens' ListImages (Maybe Natural)

The maximum number of image results returned by ListImages in paginated output. When this parameter is used, ListImages only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListImages request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListImages returns up to 100 results and a nextToken value, if applicable.

liRepositoryName :: Lens' ListImages Text

The repository whose image IDs are to be listed.

Destructuring the Response

listImagesResponse

Arguments

:: Int
lirsResponseStatus

-> ListImagesResponse

Creates a value of ListImagesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lirsImageIds - The list of image IDs for the requested repository.

lirsNextToken - The nextToken value to include in a future ListImages request. When the results of a ListImages request exceed maxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

lirsResponseStatus - -- | The response status code.

data ListImagesResponse

See: listImagesResponse smart constructor.

Response Lenses

lirsImageIds :: Lens' ListImagesResponse [ImageIdentifier]

The list of image IDs for the requested repository.

lirsNextToken :: Lens' ListImagesResponse (Maybe Text)

The nextToken value to include in a future ListImages request. When the results of a ListImages request exceed maxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

lirsResponseStatus :: Lens' ListImagesResponse Int

- | The response status code.