Network.AWS.ECR.PutImage

Description

Creates or updates the image manifest and tags associated with an image.

Creating a Request

putImage

Arguments

:: Text
piRepositoryName

-> Text
piImageManifest

-> PutImage

Creates a value of PutImage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

piRegistryId - The AWS account ID associated with the registry that contains the repository in which to put the image. If you do not specify a registry, the default registry is assumed.

piImageTag - The tag to associate with the image. This parameter is required for images that use the Docker Image Manifest V2 Schema 2 or OCI formats.

piRepositoryName - The name of the repository in which to put the image.

piImageManifest - The image manifest corresponding to the image to be uploaded.

data PutImage

See: putImage smart constructor.

Request Lenses

piRegistryId :: Lens' PutImage (Maybe Text)

The AWS account ID associated with the registry that contains the repository in which to put the image. If you do not specify a registry, the default registry is assumed.

piImageTag :: Lens' PutImage (Maybe Text)

The tag to associate with the image. This parameter is required for images that use the Docker Image Manifest V2 Schema 2 or OCI formats.

piRepositoryName :: Lens' PutImage Text

The name of the repository in which to put the image.

piImageManifest :: Lens' PutImage Text

The image manifest corresponding to the image to be uploaded.

Destructuring the Response

putImageResponse

Arguments

:: Int
pirsResponseStatus

-> PutImageResponse

Creates a value of PutImageResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pirsImage - Details of the image uploaded.

pirsResponseStatus - -- | The response status code.

data PutImageResponse

See: putImageResponse smart constructor.

Response Lenses

pirsImage :: Lens' PutImageResponse (Maybe Image)

Details of the image uploaded.

pirsResponseStatus :: Lens' PutImageResponse Int

- | The response status code.