Network.AWS.ECR.UploadLayerPart

Description

Uploads an image layer part to Amazon ECR.

Creating a Request

uploadLayerPart

Arguments

:: Text
ulpRepositoryName

-> Text
ulpUploadId

-> Natural
ulpPartFirstByte

-> Natural
ulpPartLastByte

-> ByteString
ulpLayerPartBlob

-> UploadLayerPart

Creates a value of UploadLayerPart with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ulpRegistryId - The AWS account ID associated with the registry that you are uploading layer parts to. If you do not specify a registry, the default registry is assumed.

ulpRepositoryName - The name of the repository that you are uploading layer parts to.

ulpUploadId - The upload ID from a previous InitiateLayerUpload operation to associate with the layer part upload.

ulpPartFirstByte - The integer value of the first byte of the layer part.

ulpPartLastByte - The integer value of the last byte of the layer part.

ulpLayerPartBlob - The base64-encoded layer part payload.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

data UploadLayerPart

See: uploadLayerPart smart constructor.

Request Lenses

ulpRegistryId :: Lens' UploadLayerPart (Maybe Text)

The AWS account ID associated with the registry that you are uploading layer parts to. If you do not specify a registry, the default registry is assumed.

ulpRepositoryName :: Lens' UploadLayerPart Text

The name of the repository that you are uploading layer parts to.

ulpUploadId :: Lens' UploadLayerPart Text

The upload ID from a previous InitiateLayerUpload operation to associate with the layer part upload.

ulpPartFirstByte :: Lens' UploadLayerPart Natural

The integer value of the first byte of the layer part.

ulpPartLastByte :: Lens' UploadLayerPart Natural

The integer value of the last byte of the layer part.

ulpLayerPartBlob :: Lens' UploadLayerPart ByteString

The base64-encoded layer part payload.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

Destructuring the Response

uploadLayerPartResponse

Arguments

:: Int
ulprsResponseStatus

-> UploadLayerPartResponse

Creates a value of UploadLayerPartResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ulprsRegistryId - The registry ID associated with the request.

ulprsLastByteReceived - The integer value of the last byte received in the request.

ulprsRepositoryName - The repository name associated with the request.

ulprsUploadId - The upload ID associated with the request.

ulprsResponseStatus - -- | The response status code.

data UploadLayerPartResponse

See: uploadLayerPartResponse smart constructor.

Response Lenses

ulprsRegistryId :: Lens' UploadLayerPartResponse (Maybe Text)

The registry ID associated with the request.

ulprsLastByteReceived :: Lens' UploadLayerPartResponse (Maybe Natural)

The integer value of the last byte received in the request.

ulprsRepositoryName :: Lens' UploadLayerPartResponse (Maybe Text)

The repository name associated with the request.

ulprsUploadId :: Lens' UploadLayerPartResponse (Maybe Text)

The upload ID associated with the request.

ulprsResponseStatus :: Lens' UploadLayerPartResponse Int

- | The response status code.